checkout page updates if promo is applied

+++ cinemaE-Booking/home/models.py
@@ -85,7 +85,7 @@ class CardEncr(models.Model):
 class Order(models.Model):
     orderID = models.AutoField(primary_key=True)
     total = models.IntegerField(default=0)
-    numTickets = models.IntegerField(default=0)
+    numTickets = models.FloatField(default=0)
     userID = models.ForeignKey(us, on_delete=models.CASCADE, default=1)
     showtimeID = models.ForeignKey('Showtime', on_delete=models.CASCADE, default='12/12/12 12:12')
     accountID = models.ForeignKey('Account', on_delete=models.CASCADE, default=1, blank=True, null=True)
